# src/backend/api-spec.yaml
openapi: 3.0.0
info:
  title: Hospital Network Management System API
  description: |
    API specification for the Hospital Network Management System,
    enabling secure interoperability between hospital systems.
  version: 1.0.0
servers:
  - url: https://api.hospital-network.org/v1
    description: Production API server
  - url: https://staging-api.hospital-network.org/v1
    description: Staging API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Document:
      type: object
      required:
        - patientId
        - hospitalId
        - documentType
        - title
        - filePath
        - fileSize
        - mimeType
        - md5Checksum
        - uploaderId
      properties:
        documentId:
          type: string
          format: uuid
          readOnly: true
        patientId:
          type: string
          format: uuid
        hospitalId:
          type: string
          format: uuid
        documentType:
          type: string
          enum: [mri, ct_scan, x_ray, lab_report, discharge_summary, referral, prescription, other]
        title:
          type: string
        description:
          type: string
          nullable: true
        filePath:
          type: string
        fileSize:
          type: integer
          format: int64
        mimeType:
          type: string
        md5Checksum:
          type: string
        uploadDate:
          type: string
          format: date-time
          readOnly: true
        documentDate:
          type: string
          format: date-time
          nullable: true
        uploaderId:
          type: string
          format: uuid
        version:
          type: integer
          default: 1
        tags:
          type: array
          items:
            type: string
        aiProcessed:
          type: boolean
          default: false
        aiExtractedData:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /patients/{patientId}/documents:
    post:
      summary: Upload a document for a patient
      operationId: uploadDocument
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - documentType
                - title
              properties:
                file:
                  type: string
                  format: binary
                documentType:
                  type: string
                  enum: [mri, ct_scan, x_ray, lab_report, discharge_summary, referral, prescription, other]
                title:
                  type: string
                description:
                  type: string
                documentDate:
                  type: string
                  format: date-time
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List patient documents
      operationId: listDocuments
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentType
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: hospitalId
          in: query
          schema:
            type: string
            format: uuid
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /documents/{documentId}/process:
    post:
      summary: Process a document with AI for data extraction
      operationId: processDocument
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'